# openapi.yaml
openapi: 3.1.0
info:
  title: wa-hub-demo — WhatsApp HTTP API
  version: 0.2.0
  description: >
    Self-hosted WhatsApp HTTP API (Baileys). REST + JSON, with HMAC-signed outbound webhooks.
    PRECONDITION: a WhatsApp device must be paired once (scan a QR at GET /pair) before any send
    works — otherwise send routes return 503 not_connected. Confirm readiness with
    GET /api/instance/status (connection must be "connected"). See docs/INTEGRATION.md.
  license:
    name: MIT
servers:
  - url: https://wa.yourdomain.com
    description: Your stable subdomain (recommended) — or the ephemeral https://<random>.trycloudflare.com tunnel URL
  - url: http://127.0.0.1:3060
    description: On the server itself (loopback)
security:
  - bearerAuth: []
...
paths:
...
  /api/messages/send/text:
    post:
      tags: [messages]
      summary: Send a text message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, text]
              properties:
                to: { $ref: '#/components/schemas/Recipient' }
                text: { type: string, minLength: 1, maxLength: 4096 }
                quotedMessageId: { type: string }
      responses:
        '200': { $ref: '#/components/responses/SendResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/messages/send/image:
    post:
      tags: [messages]
      summary: Send an image (by URL or base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { $ref: '#/components/schemas/Recipient' }
                imageUrl: { type: string, format: uri }
                imageBase64: { type: string, description: "base64 or data: URL, max 20 MB decoded" }
                caption: { type: string, maxLength: 1024 }
      responses:
        '200': { $ref: '#/components/responses/SendResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/messages/send/file:
    post:
      tags: [messages]
      summary: Send a document/file (by URL or base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, filename]
              properties:
                to: { $ref: '#/components/schemas/Recipient' }
                fileUrl: { type: string, format: uri }
                fileBase64: { type: string }
                filename: { type: string, minLength: 1, maxLength: 255 }
                mimetype: { type: string, default: application/octet-stream }
                caption: { type: string, maxLength: 1024 }
      responses:
        '200': { $ref: '#/components/responses/SendResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/messages/send/audio:
    post:
      tags: [messages]
      summary: Send audio / a voice note (by URL or base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { $ref: '#/components/schemas/Recipient' }
                audioUrl: { type: string, format: uri }
                audioBase64: { type: string }
                ptt: { type: boolean, default: true, description: voice note }
      responses:
        '200': { $ref: '#/components/responses/SendResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/messages/send/location:
    post:
      tags: [messages]
      summary: Send a location pin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, latitude, longitude]
              properties:
                to: { $ref: '#/components/schemas/Recipient' }
                latitude: { type: number, minimum: -90, maximum: 90 }
                longitude: { type: number, minimum: -180, maximum: 180 }
                name: { type: string, maxLength: 255 }
                address: { type: string, maxLength: 500 }
      responses:
        '200': { $ref: '#/components/responses/SendResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/messages/send/reaction:
    post:
      tags: [messages]
      summary: React to a message (empty emoji removes the reaction)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, messageId, emoji]
              properties:
                to: { $ref: '#/components/schemas/Recipient' }
                messageId: { type: string, maxLength: 255 }
                emoji: { type: string, maxLength: 8 }
                fromMe: { type: boolean, default: false }
      responses:
        '200': { $ref: '#/components/responses/SendResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/messages/markRead:
    post:
      tags: [messages]
      summary: Mark a message as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, messageId]
              properties:
                to: { $ref: '#/components/schemas/Recipient' }
                messageId: { type: string, maxLength: 255 }
                fromMe: { type: boolean, default: false }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/messages/presence:
    post:
      tags: [messages]
      summary: Send a presence/chatstate (typing, recording, online, offline) — no message is sent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, type]
              properties:
                to: { $ref: '#/components/schemas/Recipient' }
                type: { type: string, enum: [composing, recording, paused, available, unavailable] }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The HUB_TOKEN.
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Required on PUT /webhook and POST /logout only when ADMIN_TOKEN is configured.
  schemas:
...
    Recipient:
      type: string
      description: >
        +972501234567, 972501234567 (bare digits, normalized to @s.whatsapp.net),
        972501234567@s.whatsapp.net, a group <15+ digits>@g.us, or <id>@lid. No leading 0.
      example: "972501234567"
...
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
  responses:
    SendResult:
      description: Sent
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean }
              id: { type: string }
              to: { type: string }
              timestamp: { type: integer }
    Unauthorized:
      description: Missing or invalid HUB_TOKEN
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    BadRequest:
      description: >
        Invalid body / recipient / JSON — e.g. {"error":"invalid_body"} (with Zod issues) or
        {"error":"invalid_json"}. Oversized base64 media (…Base64 over the 20 MB decoded cap)
        also lands here as {"error":"invalid_body"} — "file_too_large" appears only inside
        issues[].message, NOT as the top-level error.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    TooLarge:
      description: >
        Media-by-URL (imageUrl/fileUrl/audioUrl) exceeds the 20 MB cap → {"error":"file_too_large"};
        or the whole request body exceeds 20 MB → {"error":"payload_too_large"}. (Oversized base64
        media is a 400 {"error":"invalid_body"} instead — see BadRequest.)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotConnected:
      description: WhatsApp is not paired/connected yet
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }